+++ src/maui/Fabulous.MauiControls/Views/MenuItems/MenuItem.fs
@@ -1,7 +1,5 @@
 namespace Fabulous.Maui
 
-#nowarn "0044" // Disable obsolete warnings in Fabulous.MauiControls. Please remove after deleting obsolete code.
-
 open System
 open System.IO
 open System.Runtime.CompilerServices
@@ -13,9 +11,6 @@ type IFabMenuItem =
 
 module MenuItem =
     let WidgetKey = Widgets.register<MenuItem>()
-
-    [<Obsolete("MenuItem.Accelerator is obsolete in Maui and might be removed soon. Use MenuFlyoutItem.KeyboardAccelerators instead.")>]
-    let Accelerator = Attributes.defineBindableWithEquality MenuItem.AcceleratorProperty
 
     let Clicked =
         Attributes.Mvu.defineEventNoArg "MenuItem_Clicked" (fun target -> (target :?> MenuItem).Clicked)
@@ -39,14 +34,6 @@ module MenuItemBuilders =
 
 [<Extension>]
 type MenuItemModifiers =
-    /// <summary>Set the accelerator of this widget</summary>
-    /// <param name="this">Current widget</param>
-    /// <param name="value">The accelerator value</param>
-    [<Extension>]
-    [<Obsolete("Modifier accelerator is obsolete in Maui and might be removed soon. Please use MenuFlyoutItem.keyboardAccelerators instead.")>]
-    static member inline accelerator(this: WidgetBuilder<'msg, #IFabMenuItem>, value: Accelerator) =
-        this.AddScalar(MenuItem.Accelerator.WithValue(value))
-
     /// <summary>Set the source of the icon image</summary>
     /// <param name="this">Current widget</param>
     /// <param name="value">The source of the icon image</param>